refactor: move TestServices component to customer-dashboard and update routing
